+++ cli/azd/cmd/container.go
@@ -19,11 +19,16 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/environment/azdcontext"
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
 	"github.com/azure/azure-dev/cli/azd/pkg/httputil"
+	"github.com/azure/azure-dev/cli/azd/pkg/infra/provisioning"
+	infraBicep "github.com/azure/azure-dev/cli/azd/pkg/infra/provisioning/bicep"
+	infraTerraform "github.com/azure/azure-dev/cli/azd/pkg/infra/provisioning/terraform"
 	"github.com/azure/azure-dev/cli/azd/pkg/input"
 	"github.com/azure/azure-dev/cli/azd/pkg/ioc"
 	"github.com/azure/azure-dev/cli/azd/pkg/lazy"
 	"github.com/azure/azure-dev/cli/azd/pkg/output"
+	"github.com/azure/azure-dev/cli/azd/pkg/pipeline"
 	"github.com/azure/azure-dev/cli/azd/pkg/project"
+	"github.com/azure/azure-dev/cli/azd/pkg/prompt"
 	"github.com/azure/azure-dev/cli/azd/pkg/templates"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/azcli"
 	"github.com/azure/azure-dev/cli/azd/pkg/tools/bicep"
@@ -306,6 +311,23 @@ func registerCommonDependencies(container *ioc.NestedContainer) {
 	container.RegisterSingleton(swa.NewSwaCli)
 	container.RegisterSingleton(terraform.NewTerraformCli)
 
+	// Provisioning
+	container.RegisterTransient(provisioning.NewManager)
+	container.RegisterSingleton(provisioning.NewPrincipalIdProvider)
+	container.RegisterSingleton(prompt.NewDefaultPrompter)
+
+	// Provisioning Providers
+	provisionProviderMap := map[provisioning.ProviderKind]any{
+		provisioning.Bicep:     infraBicep.NewBicepProvider,
+		provisioning.Terraform: infraTerraform.NewTerraformProvider,
+	}
+
+	for provider, constructor := range provisionProviderMap {
+		if err := container.RegisterNamedTransient(string(provider), constructor); err != nil {
+			panic(fmt.Errorf("registering IaC provider %s: %w", provider, err))
+		}
+	}
+
 	// Other
 	container.RegisterSingleton(clock.New)
 
@@ -342,6 +364,25 @@ func registerCommonDependencies(container *ioc.NestedContainer) {
 	for language, constructor := range frameworkServiceMap {
 		if err := container.RegisterNamedSingleton(string(language), constructor); err != nil {
 			panic(fmt.Errorf("registering framework service %s: %w", language, err))
+		}
+	}
+
+	// Pipelines
+	container.RegisterSingleton(pipeline.NewPipelineManager)
+	container.RegisterSingleton(func(flags *pipelineConfigFlags) *pipeline.PipelineManagerArgs {
+		return &flags.PipelineManagerArgs
+	})
+
+	pipelineProviderMap := map[string]any{
+		"github-ci":  pipeline.NewGitHubCiProvider,
+		"github-scm": pipeline.NewGitHubScmProvider,
+		"azdo-ci":    pipeline.NewAzdoCiProvider,
+		"azdo-scm":   pipeline.NewAzdoScmProvider,
+	}
+
+	for provider, constructor := range pipelineProviderMap {
+		if err := container.RegisterNamedSingleton(string(provider), constructor); err != nil {
+			panic(fmt.Errorf("registering pipeline provider %s: %w", provider, err))
 		}
 	}