+++ cli/azd/pkg/infra/provisioning/terraform/terraform_provider_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 package terraform
 
 import (
@@ -13,39 +16,23 @@ import (
 	"github.com/azure/azure-dev/cli/azd/pkg/environment"
 	"github.com/azure/azure-dev/cli/azd/pkg/exec"
 	. "github.com/azure/azure-dev/cli/azd/pkg/infra/provisioning"
+	"github.com/azure/azure-dev/cli/azd/pkg/prompt"
+	terraformTools "github.com/azure/azure-dev/cli/azd/pkg/tools/terraform"
 	"github.com/azure/azure-dev/cli/azd/test/mocks"
 
+	"github.com/azure/azure-dev/cli/azd/test/mocks/mockaccount"
+	"github.com/azure/azure-dev/cli/azd/test/mocks/mockazcli"
 	"github.com/azure/azure-dev/cli/azd/test/mocks/mockexec"
 	"github.com/stretchr/testify/require"
 )
 
 func TestTerraformPlan(t *testing.T) {
-	progressLog := []string{}
-	interactiveLog := []bool{}
-	progressDone := make(chan bool)
-
 	mockContext := mocks.NewMockContext(context.Background())
 	prepareGenericMocks(mockContext.CommandRunner)
 	preparePlanningMocks(mockContext.CommandRunner)
 
-	infraProvider := createTerraformProvider(mockContext)
-	planningTask := infraProvider.Plan(*mockContext.Context)
-
-	go func() {
-		for progressReport := range planningTask.Progress() {
-			progressLog = append(progressLog, progressReport.Message)
-		}
-		progressDone <- true
-	}()
-
-	go func() {
-		for planningInteractive := range planningTask.Interactive() {
-			interactiveLog = append(interactiveLog, planningInteractive)
-		}
-	}()
-
-	deploymentPlan, err := planningTask.Await()
-	<-progressDone
+	infraProvider := createTerraformProvider(t, mockContext)
+	deploymentPlan, err := infraProvider.Plan(*mockContext.Context)
 
 	require.Nil(t, err)
 	require.NotNil(t, deploymentPlan.Deployment)
@@ -72,16 +59,12 @@ func TestTerraformPlan(t *testing.T) {
 }
 
 func TestTerraformDeploy(t *testing.T) {
-	progressLog := []string{}
-	interactiveLog := []bool{}
-	progressDone := make(chan bool)
-
 	mockContext := mocks.NewMockContext(context.Background())
 	prepareGenericMocks(mockContext.CommandRunner)
 	preparePlanningMocks(mockContext.CommandRunner)
 	prepareDeployMocks(mockContext.CommandRunner)
 
-	infraProvider := createTerraformProvider(mockContext)
+	infraProvider := createTerraformProvider(t, mockContext)
 
 	envPath := path.Join(infraProvider.projectPath, ".azure", infraProvider.env.Dotenv()["AZURE_ENV_NAME"])
 
@@ -93,24 +76,7 @@ func TestTerraformDeploy(t *testing.T) {
 		},
 	}
 
-	deployTask := infraProvider.Deploy(*mockContext.Context, &deploymentPlan)
-
-	go func() {
-		for deployProgress := range deployTask.Progress() {
-			progressLog = append(progressLog, deployProgress.Message)
-		}
-		progressDone <- true
-	}()
-
-	go func() {
-		for deployInteractive := range deployTask.Interactive() {
-			interactiveLog = append(interactiveLog, deployInteractive)
-		}
-	}()
-
-	deployResult, err := deployTask.Await()
-	<-progressDone
-
+	deployResult, err := infraProvider.Deploy(*mockContext.Context, &deploymentPlan)
 	require.Nil(t, err)
 	require.NotNil(t, deployResult)
 
@@ -124,29 +90,9 @@ func TestTerraformDestroy(t *testing.T) {
 	preparePlanningMocks(mockContext.CommandRunner)
 	prepareDestroyMocks(mockContext.CommandRunner)
 
-	progressLog := []string{}
-	interactiveLog := []bool{}
-	progressDone := make(chan bool)
-
-	infraProvider := createTerraformProvider(mockContext)
+	infraProvider := createTerraformProvider(t, mockContext)
 	destroyOptions := NewDestroyOptions(false, false)
-	destroyTask := infraProvider.Destroy(*mockContext.Context, destroyOptions)
-
-	go func() {
-		for destroyProgress := range destroyTask.Progress() {
-			progressLog = append(progressLog, destroyProgress.Message)
-		}
-		progressDone <- true
-	}()
-
-	go func() {
-		for destroyInteractive := range destroyTask.Interactive() {
-			interactiveLog = append(interactiveLog, destroyInteractive)
-		}
-	}()
-
-	destroyResult, err := destroyTask.Await()
-	<-progressDone
+	destroyResult, err := infraProvider.Destroy(*mockContext.Context, destroyOptions)
 
 	require.Nil(t, err)
 	require.NotNil(t, destroyResult)
@@ -156,32 +102,12 @@ func TestTerraformDestroy(t *testing.T) {
 }
 
 func TestTerraformState(t *testing.T) {
-	progressLog := []string{}
-	interactiveLog := []bool{}
-	progressDone := make(chan bool)
-
 	mockContext := mocks.NewMockContext(context.Background())
 	prepareGenericMocks(mockContext.CommandRunner)
 	prepareShowMocks(mockContext.CommandRunner)
 
-	infraProvider := createTerraformProvider(mockContext)
-	getStateTask := infraProvider.State(*mockContext.Context)
-
-	go func() {
-		for progressReport := range getStateTask.Progress() {
-			progressLog = append(progressLog, progressReport.Message)
-		}
-		progressDone <- true
-	}()
-
-	go func() {
-		for deploymentInteractive := range getStateTask.Interactive() {
-			interactiveLog = append(interactiveLog, deploymentInteractive)
-		}
-	}()
-
-	getStateResult, err := getStateTask.Await()
-	<-progressDone
+	infraProvider := createTerraformProvider(t, mockContext)
+	getStateResult, err := infraProvider.State(*mockContext.Context)
 
 	require.Nil(t, err)
 	require.NotNil(t, getStateResult.State)
@@ -196,7 +122,7 @@ func TestTerraformState(t *testing.T) {
 	)
 }
 
-func createTerraformProvider(mockContext *mocks.MockContext) *TerraformProvider {
+func createTerraformProvider(t *testing.T, mockContext *mocks.MockContext) *TerraformProvider {
 	projectDir := "../../../../test/functional/testdata/samples/resourcegroupterraform"
 	options := Options{
 		Module: "main",
@@ -207,28 +133,35 @@ func createTerraformProvider(mockContext *mocks.MockContext) *TerraformProvider 
 		"AZURE_SUBSCRIPTION_ID": "00000000-0000-0000-0000-000000000000",
 	})
 
-	return NewTerraformProvider(
-		*mockContext.Context,
-		env,
-		projectDir,
-		options,
-		mockContext.Console,
-		mockContext.CommandRunner,
-		&mockCurrentPrincipal{},
-		Prompters{
-			Location: func(_ context.Context, _, _ string, _ func(loc account.Location) bool) (location string, err error) {
-				return "westus2", nil
-			},
-			Subscription: func(_ context.Context, _ string) (subscriptionId string, err error) {
-				return "00000000-0000-0000-0000-000000000000", nil
-			},
-			EnsureSubscriptionLocation: func(ctx context.Context, env *environment.Environment) error {
-				env.SetSubscriptionId("00000000-0000-0000-0000-000000000000")
-				env.SetLocation("westus2")
-				return nil
+	azCli := mockazcli.NewAzCliFromMockContext(mockContext)
+	accountManager := &mockaccount.MockAccountManager{
+		Subscriptions: []account.Subscription{
+			{
+				Id:   "00000000-0000-0000-0000-000000000000",
+				Name: "test",
 			},
 		},
+		Locations: []account.Location{
+			{
+				Name:                "location",
+				DisplayName:         "Test Location",
+				RegionalDisplayName: "(US) Test Location",
+			},
+		},
+	}
+
+	provider := NewTerraformProvider(
+		terraformTools.NewTerraformCli(mockContext.CommandRunner),
+		env,
+		mockContext.Console,
+		&mockCurrentPrincipal{},
+		prompt.NewDefaultPrompter(env, mockContext.Console, accountManager, azCli),
 	)
+
+	err := provider.Initialize(*mockContext.Context, projectDir, options)
+	require.NoError(t, err)
+
+	return provider.(*TerraformProvider)
 }
 
 func prepareGenericMocks(commandRunner *mockexec.MockCommandRunner) {